Move TrackListItem functions to helper
